# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface ContactSubmission {
  name: string;
  company: string;
  email: string;
  message: string;
  /** Honeypot field, must be empty */
  website?: string;
}

export type JobApplicationResume = {
  filename: string;
  /** Base64-encoded file content (or data URL) */
  data: string;
};

export interface JobApplication {
  name: string;
  email: string;
  phone: string;
  position: string;
  message: string;
  /** Honeypot field, must be empty */
  website?: string;
  resume?: JobApplicationResume;
}

export interface SubmissionResult {
  ok: boolean;
  emailed?: boolean;
}

export interface HealthStatus {
  status: string;
}

export type SiteContentHero = {
  headline: string;
  subheadline: string;
  ctaText: string;
  ctaLink: string;
  backgroundImage: string;
  videoUrl: string;
  showVideo: boolean;
};

export type SiteContentStatsItem = {
  value: string;
  label: string;
};

export type SiteContentAbout = {
  title: string;
  description: string;
  image: string;
  strengths?: string[];
};

export type SiteContentProductsItem = {
  id: string;
  name: string;
  category: string;
  description: string;
  features: string[];
  image: string;
  sizes?: string;
  imageFit?: string;
};

export type SiteContentCertificationsItem = {
  id: string;
  name: string;
  issuer: string;
  year: string;
  logo?: string;
};

export type SiteContentClientsItem = {
  id: string;
  name: string;
  logo: string;
};

export type SiteContentTestimonialsItem = {
  id: string;
  quote: string;
  author: string;
  company: string;
  role: string;
};

export type SiteContentContact = {
  phone: string;
  email: string;
  address: string;
  whatsapp: string;
  mapEmbed: string;
};

export type SiteContentCompany = {
  name: string;
  tagline: string;
  founded: string;
  employees: string;
  totalClients: string;
};

export type SiteContentOpeningsItem = {
  id: string;
  title: string;
  type: string;
  location: string;
  summary: string;
  requirements: string[];
  icon?: string;
  active?: boolean;
};

export type SiteContentDownloadsItem = {
  id: string;
  title: string;
  description: string;
  fileType: string;
  url: string;
};

export interface SiteContent {
  hero: SiteContentHero;
  stats: SiteContentStatsItem[];
  about: SiteContentAbout;
  products: SiteContentProductsItem[];
  certifications: SiteContentCertificationsItem[];
  clients: SiteContentClientsItem[];
  testimonials: SiteContentTestimonialsItem[];
  contact: SiteContentContact;
  company: SiteContentCompany;
  openings?: SiteContentOpeningsItem[];
  downloads?: SiteContentDownloadsItem[];
}

export interface AdminCredentials {
  password: string;
}

export interface AdminToken {
  token: string;
  message: string;
}

export interface ErrorResponse {
  error: string;
}