# lib/api-zod/src/generated/types/contactSubmission.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactSubmission {
  name: string;
  company: string;
  email: string;
  message: string;
  /** Honeypot field, must be empty */
  website?: string;
}