# lib/api-zod/src/generated/types/jobApplication.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobApplicationResume } from './jobApplicationResume';

export interface JobApplication {
  name: string;
  email: string;
  phone: string;
  position: string;
  message: string;
  /** Honeypot field, must be empty */
  website?: string;
  resume?: JobApplicationResume;
}